Rename Experience component file
